Extract content formatting into a separate class
Refs #315
