+++ lib/thredded/at_users.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 module Thredded
   class AtUsers
-    def self.render(content, post, view_context)
+    # @param users_provider [#call(usernames)] given usernames, returns a list of users.
+    def self.render(content, users_provider, view_context)
       at_names = AtNotificationExtractor.new(content).run
 
       if at_names.any?
-        members = post.readers_from_user_names(at_names)
+        members = users_provider.call(at_names)
 
         members.each do |member|
           member_path = Thredded.user_path(view_context, member)
