+++ spec/lib/thredded/post_contents.yml
