+++ spec/models/thredded/post_spec.rb
@@ -88,55 +88,4 @@ module Thredded
       expect(topic.last_user).to eq user_1
     end
   end
-
-  describe Post, '#filtered_content' do
-    let(:view_context) { ViewContextStub }
-    before(:each) { @post = build(:post) }
-    after { Thredded.user_path = nil }
-
-    module ViewContextStub
-      def main_app; end
-    end
-
-    context 'various content from post_contents.yml' do
-      yml = YAML.load(
-        File.read("#{File.dirname(__FILE__)}/post_contents.yml")
-      )
-
-      yml.each do |title, contents|
-        it "renders: '#{title}'" do
-          @post.content = contents[0]
-          expected_html = contents[1]
-
-          resulting_parsed_html = parsed_html(@post.filtered_content(view_context))
-          expected_parsed_html  = parsed_html(expected_html)
-
-          expect(resulting_parsed_html).to eq(expected_parsed_html)
-        end
-      end
-
-      def parsed_html(html)
-        Nokogiri::HTML::DocumentFragment.parse(html, &:noblanks)
-          .to_html
-          .gsub(/^\s*/, '')
-          .gsub(/\s*$/, '')
-          .gsub(/^$\n/, '')
-      end
-    end
-
-    it 'links @names of members' do
-      Thredded.user_path = ->(user) { "/whois/#{user}" }
-      post_content = 'for @sam but not @al or @kek. And @joe.'
-      sam = build_stubbed(:user, name: 'sam')
-      joe = build_stubbed(:user, name: 'joe')
-      post = build_stubbed(:post, content: post_content)
-      expected_html = '<p>for <a href="/whois/sam">@sam</a> but not @al or @kek. And <a href="/whois/joe">@joe</a>.</p>'
-
-      expect(post).to receive(:readers_from_user_names)
-        .with(%w(sam al kek joe))
-        .and_return([sam, joe])
-
-      expect(post.filtered_content(view_context)).to eq expected_html
-    end
-  end
 end
